+++ resources/views/autor.blade.php
@@ -154,7 +154,8 @@
 @section('javascript')
 {!! Html::script('js/readmore.min.js') !!}
 
-@include('components.artwork_carousel_js', ['slick_query' => '.artworks-preview'])
+{!! Html::script('js/slick.js') !!}
+{!! Html::script('js/components/artwork_carousel.js') !!}
 
 <script type="text/javascript">
     $(document).ready(function(){