+++ resources/views/katalog.blade.php
@@ -152,16 +152,7 @@
                                     list($width, $height) = getimagesize(public_path() . $item->getImagePath());
                                 @endphp
                                 <div class="ratio-box" style="padding-bottom: {{ round(($height / $width) * 100, 4) }}%;">
-    	                		<img
-                                    data-sizes="auto"
-                                    data-src="{!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'600']) !!}"
-                                    data-srcset="{!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'600']) !!} 600w,
-                                            {!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'220']) !!} 220w,
-                                            {!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'300']) !!} 300w,
-                                            {!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'600']) !!} 600w,
-                                            {!! route('dielo.nahlad', ['id' => $item->id, 'width'=>'800']) !!} 800w"
-                                    class="lazyload"
-                                    alt="{!! $item->getTitleWithAuthors() !!} ">
+    	                		     @include('components.item_image_responsive', ['item' => $item])
                                 </div>
     	                	</a>
                             <div class="item-title">