+++ resources/views/layouts/master.blade.php
@@ -37,11 +37,9 @@
 
 		{{-- JS --}}
 		<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
-		<script>
-		    document.createElement( "picture" );
-		</script>
-		{!! Html::script('js/picturefill.min.js') !!}
-        {!! Html::script('js/modernizr.custom.js') !!}
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/lazysizes/4.1.2/plugins/unveilhooks/ls.unveilhooks.min.js"></script>
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/lazysizes/4.1.4/plugins/respimg/ls.respimg.min.js"></script>
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/lazysizes/4.1.2/lazysizes.min.js"></script>
 
 		@if (App::environment() == 'production')
 		<script>
@@ -89,7 +87,6 @@
 
 	<!-- Core JavaScript Files -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.3/jquery.easing.min.js"></script>
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/lazysizes/4.0.1/lazysizes.min.js" async=""></script>
     <script src="https://unpkg.com/isotope-layout@3/dist/isotope.pkgd.min.js"></script>
 	<script src="https://unpkg.com/flickity@1.1/dist/flickity.pkgd.min.js"></script>
 	{!! Html::script('js/jquery.infinitescroll.min.js') !!}